# backend/src/utils/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.1.0",
    "title": "Bank Application",
    "contact": {
      "email": "contact@pietrzakadrian.com"
    }
  },
  "host": "bank.pietrzakadrian.com",
  "schemes": ["https", "http"],
...
  "paths": {
...
    "/api/bills/{amountMoney}/isAmountMoney": {
      "get": {
        "tags": ["Bills"],
        "summary": "Returns boolean that User has enough exist available funds",
        "description": "",
        "operationId": "isAmountMoney",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "amountMoney",
            "in": "path",
            "description": "",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "boolean",
            "schema": {
              "$ref": "#/definitions/isAmountMoney"
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/additionals/messages/isMessage": {
      "get": {
        "tags": ["Additionals"],
        "summary": "Returns boolean that user has new message",
        "description": "",
        "operationId": "isMessage",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/isMessage"
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "basicAuth": {
      "type": "basic"
    },
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
...
    "isAmountMoney": {
      "type": "object",
      "properties": {
        "isAmountMoney": {
          "type": "boolean"
        }
      }
    },
...
    "isMessage": {
      "type": "object",
      "properties": {
        "isMessage": {
          "type": "boolean"
        },
        "messageCount": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
...
  },
...
}